REFACTOR: move statics to util
